backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
langchain-groq==0.2.0
langchain==0.3.4
langchain-core==0.3.13
langgraph==0.2.40
pydantic==2.9.2